+++ Editor/Module.Utilities/DebugTools/FindReference.cs
@@ -1,9 +1,7 @@
-﻿
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Reflection;
 using System.Text;
-
 using UnityEditor;
 using UnityEngine;
 using UnityEngine.Events;