+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.1.0-preview.4] - 2020-01-16
+- added in grouping, some additional shortcuts (need to refactor to use the new shortcut system still)
+- added material / object replacement utilities
+
+## [0.1.0-preview.3] - 2020-01-13
+- version bump for openupm support
+
 ## [0.1.0-preview.1] - 2019-07-27
 - initial project setup, structure